% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billions_to_word.R
\name{billions_to_word}
\alias{billions_to_word}
\title{billions_to_word}
\usage{
billions_to_word(n)
}
\arguments{
\item{x}{Opis parametru x}
}
\value{
Opis zwracanego wyniku
}
\description{
Opis funkcji
}
